chore: move protocol pack/parse helpers to shared module
